Refactor game and agent modules for improved type safety and integration with new data structures
This commit introduces significant changes across multiple files to enhance the integration of game and agent modules with new data types. Key updates include:
- Replaced legacy imports from `game.types` and `agents.types` with structured imports from `data.types`, ensuring better organization and clarity.
- Refactored `StrategyPlanner` and `LLMAgent` to utilize the new `GameState`, `Plan`, and `PlayerPosition` classes, improving type safety and consistency in strategic planning and action execution.
- Removed outdated type definitions and classes from `game/types.py` and `agents/types.py`, streamlining the codebase.
- Enhanced tests to validate the new data structures and ensure robust functionality in strategic decision-making.

These changes collectively improve the maintainability and reliability of the code, facilitating better integration of game logic and agent behavior.
