+++ game/pre_draw.py
@@ -1,8 +1,10 @@
 from typing import List, Optional, Tuple
 
+from data.types.game_state import GameState
+from data.types.pot_types import SidePot
+
 from . import betting
 from .player import Player
-from .types import GameState, SidePot
 
 
 def handle_pre_draw_betting(
